Refactor feed component for reuse
