+++ client/main.scss
@@ -12,6 +12,7 @@
 @import "components/carousel-card/main";
 @import "icons";
 
+@import "components/feed/main";
 @import "components/fastft/main";
 
 @import "components/flexipod/main";
@@ -144,9 +145,10 @@ body .sidebar {
 	h1 {
 		@include nhFastFtTitle;
 	}
-
+}
+.feed {
 	article {
-		@include nhFastFtItem;
+		@include nhFeedItem;
 	}
 }
